# Use a  trie to speed up index construction