Fixes https://github.com/outlines-dev/outlines/issues/795

Awaiting https://github.com/outlines-dev/outlines/pull/1070 

## Problem

In `regex.py` Outlines compiles an index of legal tokens for each state of the FSM (`state_scan_tokens`)

On the `main` branch we use a naive approach:
- for each (`state`, `token`) belonging to (`FSM`, `vocabulary`), simulate FSM traversal character-by-character, saving (`state`, `token`) pairs which can successfully walk the FSM.

Calling walking the FSM `num_tokens * num_states` times is inefficient  and the current bottleneck in index construction.

## Solution

We can improve this by using a Trie. 

We know for a fact that if the token `foo` is illegal at a state, the token `foobar` is also illegal. Therefore we construct a trie, and if only if `foo` is legal, we check `foobar`, and only if `foobar` is legal, we check `foobarbaz`. 

This prevents unnecessary work from being performed and reduces runtime on average.

#### Result

Some overhead is introduced, resulting in a slight degradation in performance for simple patterns, but a great performance increase for complex patterns.
- Worst case (`ssn`): `352ms` -> `374ms`
- Best case (`complex_schema`): `3.09s` ->`1.30s`


#### Analysis
This analysis reviews the profiling details of multiple runs using `complex_schema`.

This PR is primarily an optimization of `state_scan_tokens`. Runtime of this function reduces from an average of 6.52s to 1.77 for `complex_schema`

New result profiling breakdown (performance varies by pattern):
- **`RegexGuide()`**: 4.55s
  - `create_states_mapping()`: 4.15s
    - `Pattern.to_fsm().reduce()`: 0.961s
    - `create_fsm_index_tokenizer()`: 3.34s
      - (overhead): 0.14s
      - `reduced_vocabulary()`: 0.55s
      - `create_fsm_index_end_to_end()`: 2.29s
        - `state_scan_tokens(...)`: 1.77s (359 calls)
        - (overhead): 0.47s

#### Benchmarks


(Benchmarks are relative to https://github.com/outlines-dev/outlines/pull/1070)

Benchmarks that have improved:
```
| Change   | Before [65ae1585]    | After [a8075e9c]    |   Ratio | Benchmark (Parameter)                                                                                     |
|----------|----------------------|---------------------|---------|-----------------------------------------------------------------------------------------------------------|
| -        | 3.09±0.02s           | 1.30±0.01s          |    0.42 | bench_json_schema.JsonSchemaBenchmark.time_json_schema_to_fsm('complex_schema')                           |
| -        | 1.58±0s              | 837±4ms             |    0.53 | bench_json_schema.JsonSchemaBenchmark.time_json_schema_to_fsm('simple_schema')                            |
| -        | 662±7ms              | 407±6ms             |    0.61 | bench_regex_guide.RegexGuideBenchmark.time_regex_to_guide('complex_phone')                                |
| -        | 4.23±0s              | 2.06±0s             |    0.49 | bench_regex_guide.RegexGuideBenchmark.time_regex_to_guide('complex_span_constrained_relation_extraction') |
| -        | 504±2ms              | 391±6ms             |    0.78 | bench_regex_guide.RegexGuideBenchmark.time_regex_to_guide('date')                                         |
| -        | 472±1ms              | 407±3ms             |    0.86 | bench_regex_guide.RegexGuideBenchmark.time_regex_to_guide('email')                                        |
| -        | 423±1ms              | 377±4ms             |    0.89 | bench_regex_guide.RegexGuideBenchmark.time_regex_to_guide('ip')                                           |
| -        | 571±4ms              | 475±5ms             |    0.83 | bench_regex_guide.RegexGuideBenchmark.time_regex_to_guide('url')                                          |
```

Benchmarks that have stayed the same:

<details>

```
| Change   | Before [65ae1585]    | After [a8075e9c]    |   Ratio | Benchmark (Parameter)                                                                                              |
|----------|----------------------|---------------------|---------|--------------------------------------------------------------------------------------------------------------------|
|          | 92.3±0.5μs           | 91.7±1μs            |    0.99 | bench_json_schema.JsonSchemaBenchmark.time_json_schema_to_regex('complex_schema')                                  |
|          | 50.5±0.9μs           | 50.0±0.5μs          |    0.99 | bench_json_schema.JsonSchemaBenchmark.time_json_schema_to_regex('simple_schema')                                   |
|          | 3.78±0.01s           | 4.11±0.01s          |    1.09 | bench_numba_compile.NumbaCompileBenchmark.time_compile_numba                                                       |
|          | 183±0.2μs            | 183±0.9μs           |    1    | bench_processors.LogitsProcessorPassthroughBenchmark.time_passthrough('numpy')                                     |
|          | 181±0.4μs            | 181±3μs             |    1    | bench_processors.LogitsProcessorPassthroughBenchmark.time_passthrough('torch')                                     |
|          | 247±4μs              | 247±3μs             |    1    | bench_processors.LogitsProcessorStructuredBenchmark.time_structured_generation('numpy', 'Z*')                      |
|          | 1.04±0.01ms          | 1.06±0.02ms         |    1.02 | bench_processors.LogitsProcessorStructuredBenchmark.time_structured_generation('numpy', '[^Z]*')                   |
|          | 230±0.8μs            | 224±5μs             |    0.97 | bench_processors.LogitsProcessorStructuredBenchmark.time_structured_generation('torch', 'Z*')                      |
|          | 1.03±0.01ms          | 1.04±0.01ms         |    1    | bench_processors.LogitsProcessorStructuredBenchmark.time_structured_generation('torch', '[^Z]*')                   |
|          | 17.0±0.06ms          | 17.6±0.4ms          |    1.04 | bench_regex_fsm.RegexReducedVocabularyBenchmark.time_reduced_vocabulary(10000)                                     |
|          | 178±2ms              | 175±0.7ms           |    0.98 | bench_regex_fsm.RegexReducedVocabularyBenchmark.time_reduced_vocabulary(100000)                                    |
|          | 1.87±0.02s           | 1.85±0.01s          |    0.99 | bench_regex_fsm.RegexReducedVocabularyBenchmark.time_reduced_vocabulary(1000000)                                   |
|          | 640M                 | 671M                |    1.05 | bench_regex_guide.MemoryRegexGuideBenchmark.peakmem_regex_to_guide('complex_span_constrained_relation_extraction') |
|          | 542M                 | 586M                |    1.08 | bench_regex_guide.MemoryRegexGuideBenchmark.peakmem_regex_to_guide('simple_phone')                                 |
|          | 389±0.9ms            | 384±3ms             |    0.99 | bench_regex_guide.RegexGuideBenchmark.time_regex_to_guide('simple_phone')                                          |
|          | 352±1ms              | 374±4ms             |    1.06 | bench_regex_guide.RegexGuideBenchmark.time_regex_to_guide('ssn')                                                   |
|          | 346±3ms              | 338±2ms             |    0.98 | bench_regex_guide.RegexGuideBenchmark.time_regex_to_guide('time')                                                  |
```

</details>

#### Profile

<details>

by `tottime`

```
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
      359    1.414    0.004    1.414    0.004 /home/andrew/p/outlines/outlines/fsm/regex.py:662(state_scan_tokens)
        1    1.381    1.381    4.081    4.081 /home/andrew/p/outlines/outlines/caching.py:114(wrapper)
        1    0.358    0.358    1.761    1.761 /home/andrew/p/outlines/outlines/fsm/regex.py:783(create_fsm_index_end_to_end)
    67256    0.267    0.000    0.267    0.000 {method 'index' of 'list' objects}
        1    0.231    0.231    0.231    0.231 <string>:2(ctor)
  1665873    0.205    0.000    0.205    0.000 {method 'add' of 'set' objects}
        1    0.136    0.136    1.897    1.897 /home/andrew/p/outlines/outlines/fsm/regex.py:1016(create_fsm_index_tokenizer)
      523    0.134    0.000    0.724    0.001 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/interegular/fsm.py:969(crawl)
        1    0.113    0.113    0.113    0.113 /home/andrew/p/outlines/outlines/fsm/regex.py:960(vocab_dict_to_inverted_vocab_list)
      359    0.109    0.000    0.109    0.000 {built-in method torch.tensor}
        8    0.105    0.013    0.105    0.013 {method '__reduce_ex__' of 'object' objects}
  1126880    0.101    0.000    0.101    0.000 {method 'setdefault' of 'dict' objects}
   198170    0.095    0.000    0.107    0.000 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/interegular/fsm.py:347(follow)
   114135    0.082    0.000    0.082    0.000 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/interegular/fsm.py:925(follow)
        1    0.082    0.082    0.082    0.082 /home/andrew/p/outlines/outlines/fsm/regex.py:739(get_vocabulary_transition_keys)
   128256    0.074    0.000    0.074    0.000 {method 'decode' of 'tokenizers.decoders.Decoder' objects}
   128256    0.054    0.000    0.177    0.000 /home/andrew/p/outlines/outlines/models/transformers.py:96(convert_token_to_string)
   128420    0.050    0.000    0.196    0.000 /nix/store/yd3jmjzhalskwdbcn5c6sxnsyql9a21w-python3.12-numba-0.60.0/lib/python3.12/site-packages/numba/core/typing/typeof.py:27(typeof)
```

by `cumtime`


```
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    5.395    5.395 {built-in method builtins.exec}
        1    0.000    0.000    5.395    5.395 <string>:1(<module>)
        1    0.000    0.000    5.395    5.395 /nix/store/z7xxy35k7620hs6fn6la5fg2lgklv72l-python3-3.12.4/lib/python3.12/contextlib.py:78(inner)
        1    0.029    0.029    5.394    5.394 /home/andrew/p/outlines/profile_regex_index.py:37(construct_regex_index)
        1    0.000    0.000    5.365    5.365 /home/andrew/p/outlines/outlines/fsm/guide.py:179(__init__)
        1    0.001    0.001    5.197    5.197 /home/andrew/p/outlines/outlines/caching.py:114(wrapper)
        1    0.000    0.000    5.197    5.197 /home/andrew/p/outlines/outlines/fsm/guide.py:119(create_states_mapping)
        1    0.158    0.158    4.141    4.141 /home/andrew/p/outlines/outlines/fsm/regex.py:1016(create_fsm_index_tokenizer)
        1    0.543    0.543    3.082    3.082 /home/andrew/p/outlines/outlines/fsm/regex.py:783(create_fsm_index_end_to_end)
      359    1.765    0.005    1.766    0.005 /home/andrew/p/outlines/outlines/fsm/regex.py:662(state_scan_tokens)
      523    0.178    0.000    0.957    0.002 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/interegular/fsm.py:969(crawl)
     10/1    0.001    0.000    0.911    0.911 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/interegular/patterns.py:447(to_fsm)
    128/2    0.001    0.000    0.777    0.388 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/interegular/patterns.py:453(<genexpr>)
    118/1    0.002    0.000    0.777    0.777 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/interegular/patterns.py:370(to_fsm)
        1    0.002    0.002    0.734    0.734 /home/andrew/p/outlines/outlines/fsm/regex.py:999(reduced_vocabulary)
     13/1    0.000    0.000    0.691    0.691 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/interegular/patterns.py:280(to_fsm)
      131    0.001    0.000    0.405    0.003 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/interegular/fsm.py:310(concatenate)
    67256    0.354    0.000    0.354    0.000 {method 'index' of 'list' objects}
       10    0.000    0.000    0.315    0.032 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/interegular/fsm.py:451(union)
       10    0.000    0.000    0.315    0.031 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/interegular/fsm.py:913(parallel)
        1    0.053    0.053    0.312    0.312 /home/andrew/p/outlines/outlines/fsm/regex.py:903(get_normalized_vocab)
        1    0.038    0.038    0.278    0.278 /home/andrew/p/outlines/outlines/fsm/regex.py:938(to_numba_dict)
        1    0.000    0.000    0.267    0.267 /nix/store/yd3jmjzhalskwdbcn5c6sxnsyql9a21w-python3.12-numba-0.60.0/lib/python3.12/site-packages/numba/experimental/jitclass/base.py:119(__call__)
        1    0.267    0.267    0.267    0.267 <string>:2(ctor)
  1665867    0.256    0.000    0.256    0.000 {method 'add' of 'set' objects}
   128256    0.077    0.000    0.251    0.000 /home/andrew/p/outlines/outlines/models/transformers.py:96(convert_token_to_string)
   128420    0.063    0.000    0.247    0.000 /nix/store/yd3jmjzhalskwdbcn5c6sxnsyql9a21w-python3.12-numba-0.60.0/lib/python3.12/site-packages/numba/core/typing/typeof.py:27(typeof)
   128256    0.059    0.000    0.174    0.000 /nix/store/dvqgbyv0i8bh4ddqay6cpsv9i48xi3ic-python3.12-transformers-4.43.3/lib/python3.12/site-packages/transformers/tokenization_utils_fast.py:637(convert_tokens_to_string)
        1    0.015    0.015    0.168    0.168 /home/andrew/p/outlines/outlines/fsm/guide.py:285(_cache_state_to_token_tensor)
        1    0.000    0.000    0.156    0.156 /home/andrew/p/outlines/outlines/models/transformers.py:118(__hash__)
        1    0.000    0.000    0.156    0.156 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/datasets/fingerprint.py:226(hash)
        1    0.000    0.000    0.156    0.156 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/datasets/utils/_dill.py:106(dumps)
        1    0.000    0.000    0.156    0.156 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/datasets/utils/_dill.py:101(dump)
        1    0.000    0.000    0.156    0.156 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/dill/_dill.py:418(dump)
        1    0.000    0.000    0.156    0.156 /nix/store/z7xxy35k7620hs6fn6la5fg2lgklv72l-python3-3.12.4/lib/python3.12/pickle.py:470(dump)
    179/1    0.000    0.000    0.156    0.156 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/datasets/utils/_dill.py:31(save)
    179/1    0.000    0.000    0.156    0.156 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/dill/_dill.py:367(save)
    179/1    0.000    0.000    0.155    0.155 /nix/store/z7xxy35k7620hs6fn6la5fg2lgklv72l-python3-3.12.4/lib/python3.12/pickle.py:529(save)
        1    0.000    0.000    0.155    0.155 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/datasets/utils/_dill.py:202(_save_transformersPreTrainedTokenizerBase)
      9/1    0.000    0.000    0.155    0.155 /nix/store/z7xxy35k7620hs6fn6la5fg2lgklv72l-python3-3.12.4/lib/python3.12/pickle.py:615(save_reduce)
      9/1    0.000    0.000    0.155    0.155 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/dill/_dill.py:1195(save_module_dict)
      9/1    0.000    0.000    0.155    0.155 /nix/store/z7xxy35k7620hs6fn6la5fg2lgklv72l-python3-3.12.4/lib/python3.12/pickle.py:959(save_dict)
      9/1    0.000    0.000    0.155    0.155 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/datasets/utils/_dill.py:72(_batch_setitems)
      9/1    0.000    0.000    0.155    0.155 /nix/store/z7xxy35k7620hs6fn6la5fg2lgklv72l-python3-3.12.4/lib/python3.12/pickle.py:970(_batch_setitems)
      359    0.153    0.000    0.153    0.000 {built-in method torch.tensor}
        8    0.153    0.019    0.153    0.019 {method '__reduce_ex__' of 'object' objects}
       13    0.000    0.000    0.145    0.011 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/interegular/fsm.py:364(__add__)
128744/128420    0.054    0.000    0.143    0.000 /nix/store/z7xxy35k7620hs6fn6la5fg2lgklv72l-python3-3.12.4/lib/python3.12/functools.py:904(wrapper)
   198170    0.125    0.000    0.139    0.000 /home/andrew/p/outlines/.myenv2/lib/python3.12/site-packages/interegular/fsm.py:347(follow)
        1    0.139    0.139    0.139    0.139 /home/andrew/p/outlines/outlines/fsm/regex.py:960(vocab_dict_to_inverted_vocab_list)
```

</details>