@lapp0 2024-05-15
I believe I've found a bug in `regex.py`'s `reduced_vocabulary()`

For the token `188` in the `gpt2` tokenizer (`'\x00'`), `token_tuple_np` is empty (`array([''], dtype='<U2')`), however it isn't added to `empty_token_ids`.